// backend/post-service/src/post/dto/share-post.dto.ts
import z from 'zod';
import { PostMode } from '../enums/post.enum';

export const SharePost = z.object({
  author: z.string().min(1, 'author is required'),

  caption: z.string().optional(),

  mode: z
    .enum(Object.values(PostMode) as [string, ...string[]])
    .default(PostMode.PUBLIC),

  isShare: z.boolean().default(true),

  sharePost: z.string(),
});
export type SharePostDto = z.infer<typeof SharePost>;